pushed randomness into a distribution on parameter values in the mira graph itself
